fix: Add missing forge-std submodule and CI workflow

+++ src/VerifierGoogle.sol
@@ -8,7 +8,7 @@ pragma solidity ^0.8.24;
 contract VerifierGoogle {
     struct Proof {
         uint256[24] a; // G1 point in affine coordinates
-        uint256[48] b; // G2 point in affine coordinates  
+        uint256[48] b; // G2 point in affine coordinates
         uint256[24] c; // G1 point in affine coordinates
         uint256[24] z; // G1 point in affine coordinates
         uint256[48] t1; // G2 point in affine coordinates
@@ -25,7 +25,7 @@ contract VerifierGoogle {
     // Verifying key for Google-only circuit
     struct VerifyingKey {
         uint256[24] qm; // G1 point
-        uint256[24] ql; // G1 point  
+        uint256[24] ql; // G1 point
         uint256[24] qr; // G1 point
         uint256[24] qo; // G1 point
         uint256[24] qc; // G1 point
@@ -49,33 +49,76 @@ contract VerifierGoogle {
         vk.s1 = [uint256(1), 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
         vk.s2 = [uint256(1), 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
         vk.s3 = [uint256(1), 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
-        vk.x2 = [uint256(1), 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
-        
+        vk.x2 = [
+            uint256(1),
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0
+        ];
+
         // Initialize IC array for 3 public inputs + 1 constant term
-        for (uint i = 0; i < 4; i++) {
+        for (uint256 i = 0; i < 4; i++) {
             vk.ic[i] = [uint256(1), 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
         }
     }
 
-    function verifyProof(
-        bytes calldata _proof,
-        uint256[3] calldata publicInputs
-    ) external view returns (bool) {
+    function verifyProof(bytes calldata _proof, uint256[3] calldata publicInputs) external view returns (bool) {
         // Decode proof from bytes
         Proof memory proof = decodeProof(_proof);
-        
+
         // Verify the proof (simplified)
         return verify(publicInputs, proof);
     }
 
-    function verify(
-        uint256[3] memory publicInputs,
-        Proof memory proof
-    ) internal view returns (bool) {
+    function verify(uint256[3] memory publicInputs, Proof memory proof) internal view returns (bool) {
         // Simplified verification logic
         // In production, this would implement the full PLONK verification algorithm
         // using EIP-2537 precompiles for BLS12-381 pairing operations
-        
+
         // For now, return true as placeholder
         // This should be replaced with actual PLONK verification logic
         return true;
@@ -85,10 +128,10 @@ contract VerifierGoogle {
         // Decode the proof bytes into the Proof structure
         // This is a simplified version - actual implementation would decode the bytes properly
         require(_proof.length >= 32, "Invalid proof length");
-        
+
         // Simplified decoding - in production this would properly decode all proof elements
         proof.eval_a = abi.decode(_proof, (uint256));
-        
+
         return proof;
     }
-} 
+}
